+++ CHANGES.md
@@ -18,6 +18,8 @@ New:
 - Added `video.vumeter`.
 - Added `video.slideshow`.
 - Added `video.add_text.camlimages` (#2202).
+- Added `irc.channel` operator to retrieve the contents of an IRC channel
+  (#2210).
 
 Changes:
 
@@ -44,8 +46,12 @@ Changes:
 - Made `streams_info` parameter of `output.file.hls` a record (#2173).
 - Disable scrolling by default in `video.add_text`. You can re-enable it by
   using `video.add_text(speed=70, ...)`.
+- Added "example" sections to operators documentation, we now need to populate
+  those (#2227).
 - Default implementation of `video.testsrc` is now builtin, previous
   implementation can be found under `video.testsrc.ffmpeg`.
+- Images can now generate blank audio if needed, no need to add
+  `mux_audio(audio=blank(),image)` anymore (#2230).
 
 2.0.2 (28-12-2021)
 =====